backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.0

# WebSockets & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
numpy==1.26.2
scikit-image==0.22.0

# NLP & Text Processing
spacy==3.7.2
python-dotenv==1.0.0
pandas==2.1.3

# Firebase
firebase-admin==6.2.0

# API & HTTP
requests==2.31.0
httpx==0.25.1

# Utilities
celery==5.3.4
redis==5.0.1
python-decouple==3.8
Werkzeug==3.0.1

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
pytest==7.4.3
pytest-django==4.7.0
django-debug-toolbar==4.2.0

# AWS (Optional)
boto3==1.34.1
django-storages==1.14.2

# Email & Notifications
django-anymail==10.1

# Serialization
msgpack==1.0.7

# Rate Limiting
django-ratelimit==4.1.0

# Task Scheduling
django-celery-beat==2.5.0
django-celery-results==2.5.1
